refactor: migrate source code to src directory and update project structure documentation
