--- entrypoints/popup/App.vue.bak
@@ -1,50 +1,0 @@
-<script setup lang="ts">
-import { ref, onMounted } from 'vue';
-import Login from './components/Login.vue';
-import MainView from './components/MainView.vue';
-
-const isLoggedIn = ref(false);
-const isLoading = ref(true);
-
-onMounted(async () => {
-  const result = await browser.storage.local.get(['uid', 'pwd', 'theme']);
-  if (result.theme) {
-    document.body.setAttribute('data-theme', result.theme);
-  }
-  
-  if (result.uid && result.pwd) {
-    isLoggedIn.value = true;
-  }
-  isLoading.value = false;
-});
-
-const handleLoginSuccess = () => {
-  isLoggedIn.value = true;
-};
-
-const handleLogout = async () => {
-  await browser.storage.local.remove(['uid', 'pwd']);
-  isLoggedIn.value = false;
-};
-</script>
-
-<template>
-  <div class="container">
-    <div v-if="isLoading" class="loading-screen">
-      載入中...
-    </div>
-    <Login v-else-if="!isLoggedIn" @login-success="handleLoginSuccess" />
-    <MainView v-else @logout="handleLogout" />
-  </div>
-</template>
-
-<style scoped>
-.loading-screen {
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 1.2rem;
-  font-weight: bold;
-}
-</style>
