+++ src/components/CollapsibleGuide.vue
