+++ src/components/CourseSelectorPreview.vue
