+++ src/components/Tabs.vue
