+++ src/entrypoints/assessment.content/index.ts
