+++ src/entrypoints/background.ts
