+++ src/entrypoints/course.content/course-selector.css
