+++ src/entrypoints/course.content/index.ts
