+++ src/entrypoints/file.content/index.ts
