+++ src/entrypoints/mobile/index.html
@@ -4,7 +4,8 @@
 <head>
   <meta charset="UTF-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>北科大 SSO+</title>
+  <title>北科大 SSO+ (行動版)</title>
+  <link rel="stylesheet" href="../popup/style.css">
   <link rel="preconnect" href="https://fonts.googleapis.com">
   <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
   <link
