+++ src/entrypoints/mobile/main.ts
