+++ src/entrypoints/pdf-viewer.content/index.ts
