+++ src/entrypoints/popup/App.vue
@@ -1,11 +1,11 @@
 <script setup lang="ts">
 import { browser } from 'wxt/browser';
 import { ref, onMounted } from 'vue';
-import Login from '../popup/components/Login.vue';
-import MainView from '../popup/components/MainView.vue';
-import Tabs from '../popup/components/Tabs.vue';
-import ExperimentalTab from '../popup/components/ExperimentalTab.vue';
-import Header from '../popup/components/Header.vue';
+import Login from '@/components/Login.vue';
+import MainView from '@/components/MainView.vue';
+import Tabs from '@/components/Tabs.vue';
+import ExperimentalTab from '@/components/ExperimentalTab.vue';
+import Header from '@/components/Header.vue';
 
 const isLoggedIn = ref(false);
 const isLoading = ref(true);
@@ -59,42 +59,11 @@ const handleLogout = async () => {
   </div>
 </template>
 
-<style>
-/* Global overrides for Mobile version */
-body {
-  min-width: 0 !important;
-  min-height: 0 !important;
-  overflow-x: hidden;
-}
-
-.container {
-  padding: var(--spacing-md) !important;
-  padding-bottom: 80px !important;
-}
-
-.grid-layout {
-  grid-template-columns: repeat(1, 1fr) !important;
-}
-
-.sub-cards-grid {
-  grid-template-columns: 1fr !important;
-  margin-top: 0 !important;
-}
-
-.tab-content-area {
-  padding-bottom: 120px !important;
-}
-
-/* Hide 'New Tab' and 'New Window' in mobile */
-.header-actions button.secondary {
-  display: none !important;
-}
-</style>
-
 <style scoped>
 .container {
   width: 100%;
-  min-height: 100vh;
+  height: 100%;
+  padding: var(--spacing-md);
   display: flex;
   flex-direction: column;
   background: var(--bg-sub);
@@ -108,7 +77,7 @@ body {
 }
 
 .loading-screen {
-  height: 100vh;
+  height: 100%;
   display: flex;
   flex-direction: column;
   align-items: center;
