+++ src/entrypoints/user-css.content/course-curr.css
