+++ src/entrypoints/user-css.content/course.css
