+++ src/entrypoints/user-css.content/search-curr.css
