+++ wxt.config.ts
@@ -3,6 +3,7 @@ import pkg from "./package.json";
 
 // See https://wxt.dev/api/config.html
 export default defineConfig({
+  srcDir: "src",
   outDir: "dist",
   modules: ["@wxt-dev/module-vue"],
   manifest: {
